Try again with interface in codespaces package
Revert "Extract LiveshareSession interface"

This reverts commit 592ec05ce300660e6359fbc343d22131bacb358d.
