+++ pkg/liveshare/session_test.go
@@ -46,9 +46,7 @@ func makeMockSession(opts ...livesharetest.ServerOption) (*livesharetest.Server,
 	if err != nil {
 		return nil, nil, fmt.Errorf("error connecting to Live Share: %w", err)
 	}
-
-	testSession, _ := session.(*Session)
-	return testServer, testSession, nil
+	return testServer, session, nil
 }
 
 func TestServerStartSharing(t *testing.T) {
